Made quick alteration to shooting script for aiming properly
The player would previously shoot before the camera updated. This meant that if the player aimed in one direction, then another, the previous shot would go in the first direction and not aim at the target. I added a one-frame delay between the player choosing to fire and the gun firing, to ensure the player direction had updated appropriately.